Pull out set_http_meta changes, revert unnecessary encoder changes
